requirements.txt
appdirs==1.4.4
black==21.7b0
boto3==1.20.24
botocore==1.23.24
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==5.0.0
httplib2==0.20.2
idna==3.3
jmespath==0.10.0
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.1
pathspec==0.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
six==1.16.0
tomli==1.2.1
urllib3==1.26.7
